requirements.txt
Flask
Flask-MySQL
Flask-SQLAlchemy
mysql-connector-python
